add unit debug/ex3
This is a html/css example.
If the text is in class layer1, the font color should be red.
If the text is further in class layer2, the font color should be green.
If the text is further in class layer3, the font color should be blue.
We can find that the forth 'layer2' become blue because of the missing
'</div>' tag. Further, good indentation habit makes it easy to debug, or
it would be very hard to find where is wrong.